+++ src/asp/Camera/BundleAdjustOptions.h
@@ -55,8 +55,9 @@ struct BaOptions: public asp::BaBaseOptions {
   vw::Vector2 elevation_limit;   // Expected range of elevation to limit results to.
   vw::BBox2 lon_lat_limit;       // Limit the triangulated interest points to this lonlat range
   vw::Matrix<double> initial_transform;
-  std::string fixed_cameras_indices_str, flann_method;
+  std::string fixed_cameras_indices_str, flann_method, fixed_distortion_indices_str;
   std::set<int> fixed_cameras_indices;
+  std::vector<int> fixed_distortion_indices;
   asp::IntrinsicOptions intrinsics_options;
   vw::Vector2i matches_per_tile_params;