+++ src/asp/Core/IpMatchingAlgs.cc
@@ -454,7 +454,7 @@ void listExistingMatchFiles(std::string const& prefix,
   }
 }
 
-// Given a pair of indicies, return all the matches between them.
+// Given a pair of indices, return all the matches between them.
 void matchesForPair(vw::ba::ControlNetwork const& cnet,
                     int left_cid, int right_cid,
                     std::vector<vw::ip::InterestPoint> & left_ip,